tests revised
