+++ cmake/modules/PrecompiledHeader.cmake
@@ -65,7 +65,7 @@ FUNCTION(ADD_PRECOMPILED_HEADER _targetName _input)
 	  ENDIF(FORCEINCLUDE)
 	ENDIF(_sourceWe STREQUAL ${_inputWe})
 	SET_SOURCE_FILES_PROPERTIES(${_source} PROPERTIES COMPILE_FLAGS "${PCH_COMPILE_FLAGS}")
- 	MESSAGE(STATUS "${_targetName}: SET_SOURCE_FILES_PROPERTIES(${_source} PROPERTIES COMPILE_FLAGS ${PCH_COMPILE_FLAGS} )")
+ 	#MESSAGE(STATUS "${_targetName}: SET_SOURCE_FILES_PROPERTIES(${_source} PROPERTIES COMPILE_FLAGS ${PCH_COMPILE_FLAGS} )")
       ENDIF()
     ENDFOREACH()
     IF(NOT _sourceFound)
@@ -108,7 +108,7 @@ FUNCTION(ADD_PRECOMPILED_HEADER _targetName _input)
 	GET_FILENAME_COMPONENT(_sourceWe ${_source} NAME_WE)
 	SET(PCH_COMPILE_FLAGS "${PCH_COMPILE_FLAGS} -include ${_name} -Winvalid-pch")
 	SET_SOURCE_FILES_PROPERTIES(${_source} PROPERTIES COMPILE_FLAGS "${PCH_COMPILE_FLAGS}")
-	MESSAGE(STATUS "SET_SOURCE_FILES_PROPERTIES(${_source} PROPERTIES COMPILE_FLAGS ${PCH_COMPILE_FLAGS})")
+	#MESSAGE(STATUS "SET_SOURCE_FILES_PROPERTIES(${_source} PROPERTIES COMPILE_FLAGS ${PCH_COMPILE_FLAGS})")
       ENDIF()
     ENDFOREACH()
