+++ src/result/totreemodel.cpp
@@ -35,7 +35,7 @@
 #include "result/totreemodel.h"
 
 #include "core/toconfiguration.h"
-#include "core/todatabasesetting.h"
+#include "core/todatabaseconfig.h"
 #include "core/toqvalue.h"
 #include "ts_log/ts_log_utils.h"
